% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{A}
\alias{A}
\title{Generator for Simulated Multivariate Time Series}
\description{
Coefficient matrix for a stationary simulated multivariate time series
}
\details{
Example generator matrix adapted from Table 3.2 of Gredenhoff and Karlsson (1997)
}
\references{
Gredenhoff, Mikael, and Sune Karlsson. "Lag-length selection in VAR-models using equal and unequal lag-length procedures." Computational Statistics 14.2 (1999): 171-187.
}
\author{
Will Nicholson
}
