% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{BigVAR-class}
\alias{BigVAR-class}
\title{BigVAR Object Class}
\description{
An object class to be used with cv.BigVAR
}
\details{
Construct an object of class BigVAR via the function "ConstructModel"
}
\section{Slots}{

\describe{
\item{\code{Data}}{a \eqn{T x k} multivariate time Series}

\item{\code{lagmax}}{Maximal lag order}

\item{\code{Structure}}{Penalty Structure}

\item{\code{Relaxed}}{Indicator for relaxed VAR}

\item{\code{Granularity}}{Granularity of Penalty Grid}

\item{\code{horizon}}{Desired Forecast Horizon}

\item{\code{crossval}}{Cross-Validation Procedure}

\item{\code{alpha}}{penalty for Sparse Group Lasso}

\item{\code{nseries}}{Number of Series}

\item{\code{Minnesota}}{Minnesota Prior Indicator}

\item{\code{verbose}}{Indicator for Verbose output}
}}
\seealso{
\code{\link{constructModel}}
}

