% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BigVARObjectClass.R
\docType{methods}
\name{BigVAR.est}
\alias{BigVAR.est}
\alias{BigVAR.est,BigVAR-method}
\title{BigVAR Estimation}
\usage{
BigVAR.est(object)
}
\arguments{
\item{object}{BigVAR object created from \code{ConstructModel}}
}
\value{
An array of \eqn{k \times kp} or \eqn{k\times kp+ms} coefficient matrices; one for each value of lambda.
}
\description{
Fit a BigVAR object with a structured penalty
}
\details{
Fits a using a structured penalty on a BigVAR object.  Does not perform cross-validation.
}
\examples{
data(Y)
Y=Y[1:100,]
# construct a Lasso VAR
Model1=constructModel(Y,p=4,struct="None",gran=c(50,10))
BigVAR.est(Model1)
}
\seealso{
\code{\link{constructModel}}, \code{\link{BigVAR.results}}
}

