% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BigVARObjectClass.R
\docType{methods}
\name{coef}
\alias{coef}
\alias{coef,BigVAR.results-method}
\alias{coef-methods}
\title{Default coef method BigVAR-results, returns the last coefficient matrix from the evaluation period}
\usage{
\S4method{coef}{BigVAR.results}(object)
}
\arguments{
\item{object}{BigVAR.results object created from \code{cv.BigVAR}}
}
\description{
Default coef method BigVAR-results, returns the last coefficient matrix from the evaluation period
}
\details{
displays formatted coefficient matrix
}
