% Generated by roxygen2 (4.0.2): do not edit by hand
\name{constructModel}
\alias{constructModel}
\title{Construct an object of class BigVAR}
\usage{
constructModel(Y, p, struct, gran, RVAR, h, cv, MN, verbose)
}
\arguments{
\item{Y}{T x K multivariate time series}

\item{p}{Predetermined maximal lag order}

\item{struct}{The choice of penalty structure (see details).}

\item{gran}{vector containing how deep to construct the penalty grid (parameter 1) and how many gridpoints to use (parameter 2)}

\item{RVAR}{True or False: whether to refit using the Relaxed-VAR procedure}

\item{h}{Desired forecast horizon}

\item{cv}{Cross-validation approach, either "Rolling" for rolling cross-validation or "LOO" for leave-one-out cross-validation.}

\item{MN}{Minnesota Prior Indicator}

\item{verbose,}{Verbose output while estimating}
}
\description{
Construct an object of class BigVAR
}
\details{
The choices for "struct" are as follows
\itemize{
\item{  "None" (Lasso Penalty)}
\item{  "Group" (Block Group Lasso)}
\item{  "Sparse" (Block Sparse Group Lasso)}
\item{  "Diag" (Own/Other Group Lasso) }
\item{  "SparseDiag" (Own/Other Sparse Group Lasso) }
}
}
\examples{
library(BigVAR)
data(Y)
m1=constructModel(Y,p=4,struct="None",gran=c(50,10),
RVAR=FALSE,MN=FALSE,verbose=FALSE,h=1,cv="Rolling")
}

