% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{cv.BigVAR}
\alias{cv.BigVAR}
\alias{cv.BigVAR,BigVAR-method}
\title{Cross Validation for BigVAR}
\usage{
cv.BigVAR(object)
}
\arguments{
\item{object}{BigVAR object created from \code{ConstructModel}}
}
\value{
An object of class \code{BigVAR.results}.
}
\description{
Performs rolling or leave-one-out cross-validation on a BigVAR object
}
\details{
Will perform cross validation to select penalty parameters over a training sample, then evaluate them over a test set.  Compares against sample mean, random walk, and AIC benchmarks.  The resulting object is of class \code{BigVAR.results}
}
\examples{
data(Y)
Y=Y[1:100,]
m1=constructModel(Y,p=4,struct="None",gran=c(50,10),
RVAR=FALSE,MN=FALSE,h=1,cv="Rolling",verbose=FALSE)
results=cv.BigVAR(m1)
}
\seealso{
\code{\link{constructModel}}, \code{\link{BigVAR.results}}
}

