% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FFR}
\alias{FFR}
\title{US Federal Funds Rate}
\description{
FFR Data
}
\details{
US Macroeconomic Time Series: FFR.  Ranging from Q2 1959 to Q1 2025.  Procured from FRED, Federal Reserve Bank of St. Louis.
}
\author{
Will Nicholson
}
