% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GDP}
\alias{GDP}
\title{US Gross Domestic Product}
\description{
GDP Data
}
\details{
US Macroeconomic Time Series: GDP.  Ranging from Q2 1959 to Q1 2025.  Procured from FRED, Federal Reserve Bank of St. Louis.
}
\author{
Will Nicholson
}
