% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{M1}
\alias{M1}
\title{US M1}
\description{
M1 Data
}
\details{
US Macroeconomic Time Series: M1.  Ranging from Q2 1959 to Q1 2025.  Procured from FRED, Federal Reserve Bank of St. Louis.
}
\author{
Will Nicholson
}
