% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{plot.BigVAR.results}
\alias{plot,BigVAR.results,ANY-method}
\alias{plot.BigVAR.results}
\title{Plot an object of class BigVAR.results}
\usage{
\S4method{plot}{BigVAR.results,ANY}(x, y = NULL, ...)
}
\arguments{
\item{x}{BigVAR.results object created from \code{cv.BigVAR}}

\item{y}{Needed for compatibility with generic, otherwise ignored}

\item{...}{additional arguments passed to generic}
}
\description{
Plot an object of class BigVAR.results
}
\details{
Plots optimal lambda value
}

