% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BigVARObjectClass.R
\docType{methods}
\name{plot.BigVAR}
\alias{plot.BigVAR}
\alias{plot,BigVAR-method}
\title{Plot a BigVAR object}
\usage{
\S4method{plot}{BigVAR}(x, y = NULL, ...)
}
\arguments{
\item{x}{BigVAR object created from \code{ConstructModel}}

\item{y}{NULL}

\item{...}{additional plot arguments}
}
\value{
NA, side effect is graph
}
\description{
Plot a BigVAR object
}
\details{
Uses plot.zoo to plot each individual series of \code{Y} on a single plot
}
\seealso{
\code{\link{constructModel}}
}
