% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{predict}
\alias{predict}
\alias{predict,BigVAR.results-method}
\title{Forecast using a BigVAR.results object}
\usage{
predict(object,...)
}
\arguments{
\item{object}{BigVAR.results object from \code{cv.BigVAR}}

\item{...}{additional arguments affecting the predictions produced}
}
\description{
Forecast using a BigVAR.results object
}
\details{
Provides \code{n.ahead} step forecasts using the model produced by cv.BigVAR.
}
\examples{
data(Y)
Y=Y[1:100,]
m1=constructModel(Y,p=4,struct="None",gran=c(50,10),MN=FALSE,
RVAR=FALSE,h=1,cv="Rolling",verbose=FALSE)
results=cv.BigVAR(m1)
predict(results,n.ahead=1)
}
\seealso{
\code{\link{cv.BigVAR}}
}

