% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BigVARSupportFunctions.R
\name{PredictVARX}
\alias{PredictVARX}
\title{One-step ahead predictions for VARX models}
\usage{
PredictVARX(VARXRes)
}
\arguments{
\item{VARXRes}{the results from \code{\link{VARXFit}}}
}
\value{
Returns a vector consisting of the out-of-sample forecasts for the provided \code{\link{VARXFit}} model.
}
\description{
One-step ahead predictions for VARX models
}
\examples{
data(Y)
# fit a VAR_3(3)
mod <- VARXFit(Y,3,NULL,NULL)
pred <-PredictVARX(mod)

}
\seealso{
\code{\link{VARXFit}}
}
