% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{show}
\alias{show}
\alias{show,BigVAR.results-method}
\title{Default show method for an object of class BigVAR.results}
\usage{
\S4method{show}{BigVAR.results}(object)
}
\arguments{
\item{object}{BigVAR.results object created from \code{ConstructModel}}
}
\description{
Default show method for an object of class BigVAR.results
}
\details{
prints forecast information and comparisons with mean, random walk, and AIC benchmarks
}
\seealso{
\code{\link{cv.BigVAR}}
}

