% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{show.BigVAR}
\alias{show,BigVAR-method}
\alias{show.BigVAR}
\title{Default show method for an object of class BigVAR}
\usage{
\S4method{show}{BigVAR}(object)
}
\arguments{
\item{object}{BigVAR object created from \code{ConstructModel}}
}
\value{
Data prints the first 10 rows of \code{Y}

Structure prints desired structure'

Forecast Horizon

Relaxed Indicator

Maximum lag order p
}
\description{
Default show method for an object of class BigVAR
}
\seealso{
\code{\link{constructModel}}
}

