% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BigVARObjectClass.R
\docType{methods}
\name{SparsityPlot.BigVAR.results}
\alias{SparsityPlot.BigVAR.results}
\alias{SparsityPlot.BigVAR.results,BigVAR.results-method}
\title{Sparsity Plot of a BigVAR.results object}
\usage{
SparsityPlot.BigVAR.results(object)
}
\arguments{
\item{object}{BigVAR.results object}
}
\value{
NA, side effect is graph
}
\description{
Sparsity Plot of a BigVAR.results object
}
\details{
Uses \code{levelplot} from the \code{lattice} package to plot the magnitude of each coefficient
}

