% Generated by roxygen2 (4.0.2): do not edit by hand
\name{SparsityPlot}
\alias{SparsityPlot}
\title{Sparsity Plot of a Coefficient Matrix}
\usage{
SparsityPlot(B, p, k, title = NULL)
}
\arguments{
\item{B}{\eqn{k \times kp} coefficient matrix}

\item{p}{Maximal Lag order}

\item{k}{Number of series}

\item{title}{(optional) Plot title}
}
\value{
NA, side effect is graph
}
\description{
Sparsity Plot of a Coefficient Matrix
}
\details{
Similar to SparsityPlot.BigVAR, but the input object does not need to be of class \link{BigVAR.results}
}
\examples{
data(Generator)
k=3;p=4
SparsityPlot(A[1:k,],p,k,title="Sparsity Plot of Example Generator Matrix")
}
\seealso{
\code{\link{SparsityPlot.BigVAR.results}},\code{\link{BigVAR.results}}
}

