% Generated by roxygen2 (4.0.2): do not edit by hand
\name{VarptoVar1}
\alias{VarptoVar1}
\title{Converts a VAR of order p to a VAR of order 1}
\usage{
VarptoVar1(Ai, p, k)
}
\arguments{
\item{Ai}{A list containing p k x k coefficient matrices}

\item{p}{Lag order}

\item{k}{Number of Series}
}
\value{
Returns a \eqn{kp \times kp} coefficient matrix representing all matrices contained in Ai as a VAR(1).
}
\description{
Converts a VAR of order p to a VAR of order 1
}
\examples{
library(MASS)
A1 <- matrix(c(.4,-.02,.01,-.02,.3,.02,.01,.04,.3),ncol=3,nrow=3)
A2 <- matrix(c(.2,0,0,0,.3,0,0,0,.13),ncol=3,nrow=3)
Ai=list()
Ai[[1]]=A1
Ai[[4]]=A2
A <- VarptoVar1(Ai,6,3)
}
\references{
See page 15 of Lutkepohl, "A New Introduction to Multiple Time Series Analysis"
}
\seealso{
\code{\link{MultVarSim}}
}

