% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Y}
\alias{Y}
\title{Simulated Multivariate Time Series}
\description{
Realization of a simulated multivariate time series
}
\details{
\eqn{100 \times 3} multivariate time series distributed according to the generator matrix \link{A}.
}
\author{
Will Nicholson
}
