% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Y1}
\alias{Y1}
\title{Example quarterly macroeconomic data}
\description{
Example multivariate macroeconomic time series
}
\details{
US Macroeconomic Time Series: CPI, Federal Funds Rate, Gross Domestic Product, and M1.  Ranging from Q2 1959 to Q1 2025.  Procured from FRED, Federal Reserve Bank of St. Louis.
}
\author{
Will Nicholson
}
