% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{BigVAR.results}
\alias{BigVAR.results}
\alias{BigVAR.results-class}
\title{BigVAR.results}
\description{
This class contains the results from cv.BigVAR.
}
\details{
It inherits the class BigVAR, but contains substantially more information.
}
\section{Fields}{

\describe{
\item{\code{InSampMSFE}}{In-sample MSFE from optimal value of lambda}

\item{\code{LambdaGrid}}{Grid of candidate lambda values}

\item{\code{index}}{Rank of optimal lambda value}

\item{\code{OptimalLambda}}{Value of lambda which minimizes MSFE}

\item{\code{OOSMSFE}}{Average Out of sample MSFE of BigVAR model with Optimal Lambda}

\item{\code{seoosfmsfe}}{Standard Error of Out of sample MSFE of BigVAR model with Optimal Lambda}

\item{\code{MeanMSFE}}{Average Out of sample MSFE of Unconditional Mean Forecast}

\item{\code{MeanSD}}{Standard Error of out of sample MSFE of Unconditional Mean Forecast}

\item{\code{RWMSFE}}{Average Out of sample MSFE of Random Walk Forecast}

\item{\code{RWSD}}{Standard Error of out of sample MSFE of Random Walk Forecast}

\item{\code{AICMSFE}}{Average Out of sample MSFE of AIC Forecast}

\item{\code{AICSD}}{Standard Error of out of sample MSFE of AIC Forecast}

\item{\code{betaPred}}{The final out of sample coefficient matrix of \code{B}, to be used for prediction}

\item{\code{Zvals}}{The final lagged values of \code{Y}, to be used for prediction}

\item{\code{Data}}{a \eqn{T x k} multivariate time Series}

\item{\code{lagmax}}{Maximal lag order}

\item{\code{Structure}}{Penalty Structure}

\item{\code{Relaxed}}{Indicator for relaxed VAR}

\item{\code{Granularity}}{Granularity of Penalty Grid}

\item{\code{horizon}}{Desired Forecast Horizon}

\item{\code{crossval}}{Cross-Validation Procedure}

\item{\code{alpha}}{penalty for Sparse Group Lasso}

\item{\code{nseries}}{Number of Series}

\item{\code{Minnesota}}{Minnesota Prior Indicator}

\item{\code{verbose}}{verbose indicator}
}}
\note{
One can also access any object of class BigVAR from BigVAR.results
}
\author{
Will Nicholson
}

