% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BigVARObjectClass.R
\name{constructModel}
\alias{constructModel}
\title{Construct an object of class BigVAR}
\usage{
constructModel(Y, p, struct, gran, RVAR = FALSE, h = 1, cv = "Rolling",
  MN = FALSE, verbose = TRUE, IC = TRUE, VARX = list(),
  T1 = floor(nrow(Y)/3), T2 = floor(2 * nrow(Y)/3), ONESE = FALSE)
}
\arguments{
\item{Y}{\eqn{T x k} multivariate time series or Y \eqn{T x k+m} endogenous and exogenous series, respectively}

\item{p}{Predetermined maximal lag order (for modeled series)}

\item{struct}{The choice of penalty structure (see details).}

\item{gran}{vector containing how deep to construct the penalty grid (parameter 1) and how many gridpoints to use (parameter 2)}

\item{RVAR}{True or False: whether to refit using the Relaxed-VAR procedure}

\item{h}{Desired forecast horizon}

\item{cv}{Cross-validation approach, either "Rolling" for rolling cross-validation or "LOO" for leave-one-out cross-validation.}

\item{MN}{Minnesota Prior Indicator}

\item{verbose}{Verbose output while estimating}

\item{IC}{True or False: whether to include AIC and BIC benchmarks}

\item{VARX}{List containing VARX model specifications.}

\item{T1}{Index of time series in which to start cross validation}

\item{T2}{Index of times series in which to start forecast evaluation}

\item{ONESE}{True or False: whether to use the "One Standard Error Heuristic"}
}
\description{
Construct an object of class BigVAR
}
\details{
The choices for "struct" are as follows
\itemize{
\item{  "None" (Lasso Penalty)}
\item{  "Lag" (Lag Group Lasso)}
\item{  "SparseLag" (Lag Sparse Group Lasso)}
\item{  "Diag" (Own/Other Group Lasso) }
\item{  "SparseDiag" (Own/Other Sparse Group Lasso) }
\item{  "EFX" (Endogenous First VARX)}
\item{  "HVARC" (Componentwise Hierarchical Group Lasso) }
\item{  "HVAROO" (Own/Other Hierarchical Group Lasso) }
\item{  "HVARELEM" (Elementwise Hierarchical Group Lasso)}
\item{  "Tapered" (Lag weighted Lasso)}
}

VARX specifications consist of a list with entry k1 denoting the series that are to be modeled and entry s to denote the maximal lag order for exogenous series.
}
\note{
The specifications "None", "Lag," "SparseLag," "SparseDiag," and "Diag" can accomodate both VAR and VARX models.  EFX only applies to VARX models.  "HVARC," "HVAROO," "HVARELEM," and "Tapered" can only be used with VAR models.
}
\examples{
library(BigVAR)
# VARX Example
VARX=list()
VARX$k=2 # indicates that the first two series are modeled
VARX$s=2 # sets 2 as the maximal lag order for exogenous series
data(Y)
T1=floor(nrow(Y)/3)
T2=floor(2*nrow(Y)/3)
Model1=constructModel(Y,p=4,struct="None",gran=c(50,10),verbose=FALSE,VARX=VARX,T1=T1,T2=T2)
}
\references{
William B Nicholson, Jacob Bien, and David S Matteson. "Hierarchical vector autoregression." arXiv preprint 1412.5250, 2014.

William B Nicholson, David S. Matteson, and Jacob Bien (2015), "Structured regularization for large vector
autoregressions with exogenous variables," \url{http://www.wbnicholson.com/Nicholsonetal2015.pdf}.
}

