% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BigVARObjectClass.R
\docType{methods}
\name{cv.BigVAR}
\alias{cv.BigVAR}
\alias{cv.BigVAR,BigVAR-method}
\title{Cross Validation for BigVAR}
\usage{
cv.BigVAR(object)
}
\arguments{
\item{object}{BigVAR object created from \code{ConstructModel}}
}
\value{
An object of class \code{BigVAR.results}.
}
\description{
Performs rolling cross-validation on a BigVAR object
}
\details{
Will perform cross validation to select penalty parameters over a training sample, then evaluate them over a test set.  Compares against sample mean, random walk, AIC, and BIC benchmarks.  The resulting object is of class \code{BigVAR.results}
}
\examples{
data(Y)
Y=Y[1:100,]
# construct a Lasso VAR
Model1=constructModel(Y,p=4,struct="None",gran=c(50,10))
results=cv.BigVAR(Model1)
}
\seealso{
\code{\link{constructModel}}, \code{\link{BigVAR.results}}
}

