% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BigVARObjectClass.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,BigVAR.results-method}
\title{Default show method for an object of class BigVAR.results}
\usage{
\S4method{show}{BigVAR.results}(object)
}
\arguments{
\item{object}{BigVAR.results object created from \code{cv.BigVAR}}
}
\description{
Default show method for an object of class BigVAR.results
}
\details{
prints forecast results and additional diagnostic information as well as comparisons with mean, random walk, and AIC, and BIC benchmarks
}
\seealso{
\code{\link{cv.BigVAR}},\code{\link{BigVAR.results}}
}
