% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BigVARObjectClass.R
\docType{methods}
\name{show.BigVAR}
\alias{show.BigVAR}
\alias{show,BigVAR-method}
\title{Default show method for an object of class BigVAR}
\usage{
\S4method{show}{BigVAR}(object)
}
\arguments{
\item{object}{\code{BigVAR} object created from \code{ConstructModel}}
}
\value{
Displays the following information about the BigVAR object:
\itemize{
\item{Prints the first 5 rows of \code{Y}}
\item{ Penalty Structure}
\item{ Relaxed Least Squares Indicator}
\item{Maximum lag order} 
\item{ VARX Specifications (if applicable)}
\item{Start, end of cross validation period}
}
}
\description{
Default show method for an object of class BigVAR
}
\seealso{
\code{\link{constructModel}}
}
