% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{SparsityPlot.BigVAR.results}
\alias{SparsityPlot.BigVAR.results}
\alias{SparsityPlot.BigVAR.results,BigVAR.results-method}
\title{Sparsity Plot of a BigVAR.results object}
\usage{
SparsityPlot.BigVAR.results(object)
}
\arguments{
\item{object}{BigVAR.results object}
}
\value{
NA, side effect is graph
}
\description{
Sparsity Plot of a BigVAR.results object
}
\details{
Uses levelplot from the lattice package to plot the magnitude of each coefficient
}
\seealso{
\code{\link{SparsityPlot}}
}

