% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BigVARObjectClass.R
\docType{methods}
\name{SparsityPlot.BigVAR.results}
\alias{SparsityPlot.BigVAR.results}
\alias{SparsityPlot.BigVAR.results,BigVAR.results-method}
\title{Sparsity Plot of a BigVAR.results object}
\usage{
SparsityPlot.BigVAR.results(object)
}
\arguments{
\item{object}{BigVAR.results object}
}
\value{
NA, side effect is graph
}
\description{
Sparsity Plot of a BigVAR.results object
}
\details{
Uses \code{levelplot} from the \code{lattice} package to plot the magnitude of each coefficient in the last coefficient estimated by \code{cv.BigVAR}.
}
\examples{
data(Y)
Y <- Y[1:100,]
Model1 <- constructModel(Y,p=4,struct="Basic",gran=c(50,10),verbose=FALSE)
SparsityPlot.BigVAR.results(cv.BigVAR(Model1))
}
\seealso{
\code{\link{cv.BigVAR}}, \code{\link{BigVAR.results}}
}
